Added simple signal-slot template class
This was noted in earlier commit message; there is a lot of repeated
code that iterates over a vector of callbacks in various classes, so it
makes sense to abstract this out.

Could not resist playing with variadic templates, forwarding references,
etc etc (against my better judgement), so.. there we go.  Hope it's not
too horribly written (this was "learning/referesher" exercise, so should
probably review code again at later time).

Not yet incorporated in the rest of the code, but that should be easy.