Use new sigslot class everywhere (except mqtt)
Code now uses the signal-slot template class, rather than repeating
add-callback/iterate-and-invoke code each time.

Still undecided if I want to use signals in mqtt topic callbacks (it's
already a hashmap, and it would become a hashmap-of-vectors, but I don't
expect to ever need multiple callbacks on one topic and this is an MCU
after all..).  Furthermore, the potential un-safety(?) of connection_t IDs
(see comments in sigslot.h) might become a real concern (since, if trivally
implemented, a connection ID returned by an MQTT object would *not* be
unique even within that object's context...).  Of course, an easy
solution would be to not expose disconnect from MQTT, but need to think
about this a little more carefully, perhaps.

So, left MQTT callbacks as-is, at least for now.